scripts/permutation-analysis/pyproject.toml
[tool.flake8]
max-line-length = 100
ignore = ["BLK999", "W503", "PT009"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
lint.select = ["E4", "E7", "E9", "F", "B", "E501"]
lint.extend-select = ["I"]

[tool.ruff.format]
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
addopts = "-ra -v --strict-markers -m 'not manual_test'"
markers = [
    "unit_test: Tests the behavior of one function (unit) in isolation from external infrastructure.",
    "integration_test: Tests the behavior of one function that integrates with external infrastructure.",
    "end_to_end_test: Tests the behavior of a full application, end-to-end.",
    "manual_test: Marks a test as a manual test that should not be run automatically."
]
testpaths = [
    "tests"
]

[tool.coverage.run]
source = [
    "permutation_analysis"
]

[tool.poetry]
name = "permutation-analysis"
version = "0.1.0"
description = "Package to analyze tennis match scores permutations"
authors = ["John Larkin <john@johnjlarkin.com>"]
readme = "README.md"
packages = [{include = "permutation_analysis"}]

[tool.poetry.dependencies]
python = "^3.12"
supabase = "^2.9.1"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}
tqdm = "^4.66.5"
types-tqdm = "^4.66.0.20240417"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
